Splitting out DirPath
