+++ mundane-io/src/main/scala/com/ambiata/mundane/io/FilePath.scala
@@ -29,31 +29,3 @@ object FilePath {
   def unsafe(s: String): FilePath =
     fromString(s).getOrElse(sys.error("FilePath.unsafe on an invalid string."))
 }
-
-object DirPath {
-  def Root: DirPath =
-    com.ambiata.mundane.path.Root
-
-  def Relative: DirPath =
-    com.ambiata.mundane.path.Relative
-
-  def fromFile(f: File): DirPath =
-    unsafe(f.getPath)
-
-  def fromString(s: String): Option[DirPath] = {
-    s.split("/").toList match {
-      case "" :: Nil =>
-        None
-      case "" :: parts =>
-        parts.traverse(FileName.create).map(FilePath.fromList(Root, _))
-      case parts =>
-        parts.traverse(FileName.create).map(FilePath.fromList(Relative, _))
-    }
-  }
-
-  def unsafe(s: String): DirPath =
-    fromString(s).getOrElse(sys.error("DirPath.unsafe on an invalid string."))
-
-  def fromURI(s: URI): Option[DirPath] =
-    fromString(s.getPath)
-}
